Added BridgeHue class
Needed a sensible way of getting a new API key or "username" to use with
the bridge but overloading the constructor in HueLight seemed illogical
as it isn't a light object we want. HueBridge has one method,
GetNewAPIKey() which either returns the error message spat out by the
Bridge if you forget to press the link button or, should (I haven'
tested it yet) spit out the username/API key/UserID. Ideally, you should
record this for later use rather than always going to the trouble of
creating a new username. i provide this class as more of a helper than
anything else.