Enum use
I was going to declare an Enum for the light state but decided it was
going to need some more attention than leaving it as a bool. It already
is a bool in the REST API so might as well stay as one.
Updated readme to clarify use of the calss and it can be dotsourced via
Import-Module.