Added some error checking
Added a way to ignore/throw an error when setting of the Colour
Temperature setting for lights that do not support this (Bloom). Added
error handling (throws) to setting of CT, HSB methods to enable use of
Try Catch in your own scripts. Updating Using-PoSHue.ps1 to demonstrate
use of Try Catch blocks.