BUG: RGBtoXYZ returning incorrect Y value
The RGB to XYZ method was returning incorrect values for Y - I thought
it was just a consequence of smoothed values, it wasn't. I was using a
value, overwriting it and then using the overwritten value in the Y
calculation meaning it was always incorrect - sometimes not by much,
others by vast amounts. I've now resolved this bug and the XYZ values
are bang on per Gamut. Hooray.